Refactor code by moving 'ls' function to common.io module
The 'ls' function, which lists all files in a directory and its subdirectories, has been moved from dcmcp.py to a new file io.py within the rtpytools/common directory. This change improves modularization and usability of the code across the project.
